#1503 renamed WithSparkDf to SparkDfGenMod
