+++ src/main/python/smv/smvmodule.py
@@ -46,18 +46,18 @@ class SmvOutput(object):
         return None
 
 
-class WithSparkDf(SmvGenericModule):
+class SparkDfGenMod(SmvGenericModule):
     """Base class for SmvModules create Spark DFs
     """
 
     def __init__(self, smvApp):
-        super(WithSparkDf, self).__init__(smvApp)
+        super(SparkDfGenMod, self).__init__(smvApp)
         self.dqmTimeElapsed = None
 
     #########################################################################
     # User interface methods
     #
-    # WithSparkDf specific:
+    # SparkDfGenMod specific:
     # - dqm: Optional, default SmvDQM()
     #########################################################################
     def dqm(self):
@@ -100,13 +100,13 @@ class WithSparkDf(SmvGenericModule):
 
     # Override this method to add the edd calculation if config
     def _calculate_user_meta(self):
-        super(WithSparkDf, self)._calculate_user_meta()
+        super(SparkDfGenMod, self)._calculate_user_meta()
         if (self.smvApp.py_smvconf.force_edd()):
             self.calculate_edd()
 
     # Override this method to add the dqmTimeElapsed
     def _finalize_meta(self):
-        super(WithSparkDf, self)._finalize_meta()
+        super(SparkDfGenMod, self)._finalize_meta()
         self.module_meta.addSchemaMetadata(self.data)
         # Need to add duration at the very end, just before persist
         self.module_meta.addDuration("dqm", self.dqmTimeElapsed)
@@ -115,7 +115,7 @@ class WithSparkDf(SmvGenericModule):
     def _do_action_on_df(self, func, df, desc):
         name = self.fqn()
         self.smvApp.sc.setJobGroup(groupId=name, description=desc)
-        (res, secondsElapsed) = super(WithSparkDf, self)._do_action_on_df(func, df, desc)
+        (res, secondsElapsed) = super(SparkDfGenMod, self)._do_action_on_df(func, df, desc)
 
         # Python api does not have clearJobGroup
         # set groupId and description to None is equivalent
@@ -152,7 +152,7 @@ class WithSparkDf(SmvGenericModule):
         self.module_meta.addDqmValidationResult(validation_result.toJSON())
 
 
-class SmvSparkDfModule(SmvProcessModule, WithSparkDf):
+class SmvSparkDfModule(SmvProcessModule, SparkDfGenMod):
     """Base class for SmvModules create Spark DFs
     """
 
@@ -199,27 +199,6 @@ class SmvSparkDfModule(SmvProcessModule, WithSparkDf):
         """
         return None
 
-
-    @abc.abstractmethod
-    def run(self, i):
-        """User-specified definition of the operations of this SmvModule
-
-            Override this method to define the output of this module, given a map
-            'i' from input SmvGenericModule to resulting DataFrame. 'i' will have a
-            mapping for each SmvGenericModule listed in requiresDS. E.g.
-
-            def requiresDS(self):
-                return [MyDependency]
-
-            def run(self, i):
-                return i[MyDependency].select("importantColumn")
-
-            Args:
-                (RunParams): mapping from input SmvGenericModule to DataFrame
-
-            Returns:
-                (DataFrame): output of this SmvModule
-        """
 
 
     # All publish related methods should be moved to generic output module class
